requirements.txt
flask
pypdf
sentence-transformers
faiss-cpu
ctransformers
numpy
